--- a/.github/workflows/specs.yml
+++ b/.github/workflows/specs.yml
@@ -7,8 +7,8 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        ruby: ['2.7', '3.0', '3.1' ]
-        redis: ['6.2', '7.0']
+        ruby: ['3.0', '3.1', '3.2']
+        redis: ['6.2', '7.2']
     name: Ruby ${{ matrix.ruby }}, Redis ${{ matrix.redis }}
 
     services: