--- a/Gemfile
+++ b/Gemfile
@@ -3,3 +3,17 @@
 source 'https://rubygems.org'
 
 gemspec
+
+group :development do
+  gem 'benchmark-ips', '~> 2.3'
+  gem 'database_cleaner-redis', '~> 2.0'
+  gem 'debug'
+  gem 'fakeredis', '~> 0.8'
+  gem 'rake', '~> 13.0'
+  gem 'redis', '~> 4.1'
+  gem 'rspec', '~> 3.11'
+  gem 'rubocop', '~> 1.0.0'
+  gem 'simplecov', '~> 0.21'
+  gem 'simplecov-lcov', '~> 0.8'
+  gem 'timecop', '~> 0.9'
+end