--- a/lib/stoplight/version.rb
+++ b/lib/stoplight/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Stoplight
-  VERSION = Gem::Version.new('3.0.1')
+  VERSION = Gem::Version.new('4.0.0')
 end